+++ src/main/java/com/simsimbookstore/frontserver/book/controller/BookController.java
@@ -1,0 +1,4 @@
+package com.simsimbookstore.frontserver.book.controller;
+
+public class BookController {
+}
