--- src/main/java/com/simsimbookstore/frontserver/controller/RestController.java
@@ -1,4 +1,0 @@
-package com.simsimbookstore.frontserver.controller;
-
-public class RestController {
-}
