--- src/main/java/com/simsimbookstore/frontserver/request/SocialUserRequest.java
@@ -1,4 +1,0 @@
-package com.simsimbookstore.frontserver.request;
-
-public class SocialUserRequest {
-}
