+++ src/main/java/com/simsimbookstore/frontserver/user/controller/RestController.java
@@ -1,0 +1,4 @@
+package com.simsimbookstore.frontserver.user.controller;
+
+public class RestController {
+}
