+++ src/main/java/com/simsimbookstore/frontserver/user/feign/PaycoAuthServiceClient.java
@@ -1,4 +1,4 @@
-package com.simsimbookstore.frontserver.feign;
+package com.simsimbookstore.frontserver.user.feign;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
