+++ src/main/java/com/simsimbookstore/frontserver/user/request/SocialUserRequest.java
@@ -1,0 +1,4 @@
+package com.simsimbookstore.frontserver.user.request;
+
+public class SocialUserRequest {
+}
