+++ src/main/java/com/simsimbookstore/frontserver/user/service/CustomOauth2UserService.java
@@ -1,4 +1,4 @@
-package com.simsimbookstore.frontserver.service;
+package com.simsimbookstore.frontserver.user.service;
 
 import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
