+++ src/main/java/com/simsimbookstore/frontserver/user/service/CustomUserDetailsService.java
@@ -1,6 +1,6 @@
-package com.simsimbookstore.frontserver.service;
+package com.simsimbookstore.frontserver.user.service;
 
-import com.simsimbookstore.frontserver.request.LocalUserRequest;
+import com.simsimbookstore.frontserver.user.request.LocalUserRequest;
 import com.simsimbookstore.frontserver.security.userDetails.CustomUserDetails;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.security.core.userdetails.UserDetails;
