+++ src/main/java/com/simsimbookstore/frontserver/user/service/SocialService.java
@@ -1,4 +1,4 @@
-package com.simsimbookstore.frontserver.service;
+package com.simsimbookstore.frontserver.user.service;
 
 public interface SocialService {
