+++ src/main/resources/templates/index.html
@@ -10,5 +10,6 @@
   <a href="/login">로그인 폼</a> <br><br>
 
   <a href="/users/myPage">마이페이지</a><br><br>
+  <a href="/books">도서 목록 불러오기</a><br><br>
 </body>
 </html>
