+++ src/main/resources/templates/userRegisterForm.html
@@ -18,20 +18,20 @@
   <input type="password" id="password" name="password" required><br><br>
 
   <!-- Name -->
-  <label for="name">Name:</label>
-  <input type="text" id="name" name="name" required><br><br>
+  <label for="userName">Name:</label>
+  <input type="text" id="userName" name="userName" required><br><br>
 
   <!-- Mobile Number -->
-  <label for="mobile_number">Mobile Number:</label>
-  <input type="tel" id="mobile_number" name="mobile_number" pattern="[0-9]{10,15}" required><br><br>
+  <label for="mobileNumber">Mobile Number:</label>
+  <input type="tel" id="mobileNumber" name="mobileNumber" pattern="[0-9]{10,15}" required><br><br>
 
   <!-- Email -->
   <label for="email">Email:</label>
   <input type="email" id="email" name="email" required><br><br>
 
   <!-- Birthday -->
-  <label for="birthday">Birthday:</label>
-  <input type="date" id="birthday" name="birthday" required><br><br>
+  <label for="birth">Birthday:</label>
+  <input type="date" id="birth" name="birth" required><br><br>
 
   <!-- Gender -->
   <label for="gender">Gender:</label>
